\name{cigandalc.dat}
\alias{cigandalc.dat}
\docType{data}
\title{ Raw data on cigarette and alcohol use}
\encoding{latin1}
\description{
  Values from 1204 subjects on 8 variables related to cigarette and alcohol use
}
\usage{data(cigandalc.dat)}
\format{
  A data frame with 1204 observations on 8 variables.
}
\source{
  Bentler, P. M. (2004). EQS 6 structural equations program manual. Encino, CA: Multivariate Software.
}
\references{
  Data originally collected by Duncan, T. E., Duncan, S. C., Alpert, A., Hops, H., Stoolmiller, M., & Muthén, B. (1997). 
  Latent variable modeling of longitudinal and multilevel substance use data. Multivariate Behavioral Research, 32, 275-318.
  For its use in the illustration see also Levy, R., & Hancock, G. R. (2007). A framework of statistical tests 
  for comparing mean and covariance structure models. Multivariate Behavioral Research, 42, 33-36.
}
\keyword{datasets}
