\name{model.A.df}
\alias{model.A.df}
\docType{data}
\title{ Model A's degrees of freedom}
\encoding{latin1}
\description{
  The number 25, the degrees of freedom for Model A in the illustration in Levy, R., & Hancock, G. R. (2007). A framework of statistical tests 
  for comparing mean and covariance structure models. Multivariate Behavioral Research, 42, 33-36.
}
\usage{data(model.A.df)}
\source{
  Levy, R., & Hancock, G. R. (2007). A framework of statistical tests 
  for comparing mean and covariance structure models. Multivariate Behavioral Research, 42, 33-36.
}
\references{
  Data and models based off of original data collection and model formulation by Duncan, T. E., Duncan, S. C., Alpert, A., Hops, H., Stoolmiller, M., & Muthén, B. (1997). 
  Latent variable modeling of longitudinal and multilevel substance use data. Multivariate Behavioral Research, 32, 275-318.
}
\keyword{datasets}
