\name{model.B.cov.matrix}
\alias{model.B.cov.matrix}
\docType{data}
\title{ Model B's model-implied covariance matrix}
\encoding{latin1}
\description{
  An 8x8 matrix: the model-implied covariance matrix from Model B in the illustration in Levy, R., & Hancock, G. R. (2007). A framework of statistical tests 
  for comparing mean and covariance structure models. Multivariate Behavioral Research, 42, 33-36.
}
\usage{data(model.B.cov.matrix)}
\source{
  Levy, R., & Hancock, G. R. (2007). A framework of statistical tests 
  for comparing mean and covariance structure models. Multivariate Behavioral Research, 42, 33-36.
}
\references{
  Data and models based off of original data collection and model formulation by Duncan, T. E., Duncan, S. C., Alpert, A., Hops, H., Stoolmiller, M., & Muthén, B. (1997). 
  Latent variable modeling of longitudinal and multilevel substance use data. Multivariate Behavioral Research, 32, 275-318.
}
\keyword{datasets}