\name{SEMModComp-package}
\alias{SEMModComp-package}
\alias{SEMModComp}
\docType{package}
\title{
Conduct tests of difference in fit for mean and covariance structure models as in structural equation modeling (SEM)
}
\description{
Chi-squared and normal theory likelihood ratio tests for mean and covariance structure models as in structural equation modeling (SEM).  
Used to statistically compare models in accordance with the framework described by Levy and Hancock (2007).
}
\details{
\tabular{ll}{
Package: \tab SEMModComp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-02-16\cr
License: \tab GPL (version 3 or later)
\cr
LazyLoad: \tab yes\cr
}
}
\author{
Roy Levy <Roy.Levy@asu.edu>
}
\references{
Levy, R., & Hancock, G. R. (2007). A framework of statistical tests for comparing mean and covariance structure models. Multivariate Behavioral Research, 42, 33-36.
Further documentation available at http://www.public.asu.edu/~rlevy2/papers&software.html
}
